script: replace cut with read
Executing 'cut' five times per line makes running script quite slow.
Patch replaces 'cut' with builtin 'read' that splits string to a number
of variables. This speeds up execution of litclock for three times for a
worst case.

Path to a shell in shebang has been updated: dash(1) on Debian is not
compatible with POSIX shell, for example it doesn't support redirection.
Default shell in /bin/sh is replaced with path to Bash. However,
litclock script is compatible with Korn shell, that is actually POSIX
compatible.